f(app/rute login user) the rute was create  to log user en db
