+++ api/V1/routes/usersRouter.js
@@ -1,9 +1,13 @@
+const express=require("express"),
+      router=express.Router(),
+      {jwtCheck,checkpermissions}=require('../../middleware/jwtLoginUser');
+
 const { getUserLogin } = require("../../controllers/usersController");
 
-const express=require("express"),
-      router=express.Router();
+
+
 
 router
-      .get('/', getUserLogin) 
+      .get('/',jwtCheck,getUserLogin) 
 
 module.exports=router;
