+++ api/app.js
@@ -4,9 +4,7 @@ const express=require("express"),
       cors=require('cors'),// providing a Connect/Express middleware that can be used to enable CORS with various options.
       v1ProducstRouter=require('./V1/routes/allProductsRouter'),
       v1ProductCreate=require('./V1/routes/CreateProductRouter'),
-      v1UseRollRoll=require('./V1/routes/userRollRouter'),
-      v1UsersRouter=require('./v1/routes/usersRouter'),
-      {jwtCheck,checkpermissions}=require('./middleware/jwtLoginUser');
+      v1UsersRouter=require('./v1/routes/usersRouter');
    
 
 
@@ -17,19 +15,15 @@ app.use(cors())
 
 //middlewere JSON WEB TOKEN
 
-
+//ruta general  sin registro
 app.use("/products",v1ProducstRouter);
 
+//ruta para usuarios registrados
+app.use("/loginUsers",v1UsersRouter);
+
+
+///proximamente ruta para roll admi
 app.use("/create",v1ProductCreate)
 
-
-
-
-
-//app.use(jwtCheck)
-//rote is all relation with users login,register
-//app.use("/users",v1UsersRouter)
-
-app.use("/admidrol",jwtCheck,checkpermissions,v1UseRollRoll)
 
 module.exports=app
