+++ api/controllers/usersController.js
@@ -1,12 +1,12 @@
-const { getLoginResponse } = require("../service/users/getLogin")
+const { getLoginResponse } = require("../service/users/getLoginService")
 
 
 const getUserLogin=async(req,res)=>{
 
     try {
-        const allproducts=await getLoginResponse(req)
+        const getLoginUser=await getLoginResponse(req)
     
-        res.status(200).json(allproducts)
+        res.status(200).json(getLoginUser)
     } catch (error) {
         res.status(500).json({msg:error.message})
     }
