--- api/service/users/getLogin.js
@@ -1,16 +1,0 @@
-const axios=require('axios')
-
-const getLoginResponse =async(req)=>{
-   
-    const accessToken=req.headers.authorization.split(' ')[1];
-    console.log(accessToken)
-    const response=await axios.get(`https://dev-nzbce16c.us.auth0.com/userinfo`,{headers:{
-          authorization:`Bearer ${accessToken}`
-    }})
-    const userInfo=response.data
-    return userInfo
-}
-
-module.exports = {
-    getLoginResponse
-}
